+++ jabgui/src/main/java/org/jabref/gui/preview/PreviewPreferences.java
@@ -29,7 +29,14 @@ public class PreviewPreferences {
 
     private final BooleanProperty shouldDownloadCovers = new SimpleBooleanProperty();
 
-    public PreviewPreferences(List<PreviewLayout> layoutCycle, int layoutCyclePosition, TextBasedPreviewLayout customPreviewLayout, String defaultCustomPreviewLayout, boolean showPreviewAsExtraTab, boolean showPreviewEntryTableTooltip, List<Path> bstPreviewLayoutPaths, boolean shouldDownloadCovers) {
+    public PreviewPreferences(List<PreviewLayout> layoutCycle,
+                              int layoutCyclePosition,
+                              TextBasedPreviewLayout customPreviewLayout,
+                              String defaultCustomPreviewLayout,
+                              boolean showPreviewAsExtraTab,
+                              boolean showPreviewEntryTableTooltip,
+                              List<Path> bstPreviewLayoutPaths,
+                              boolean shouldDownloadCovers) {
         this.layoutCycle = FXCollections.observableArrayList(layoutCycle);
         this.layoutCyclePosition = new SimpleIntegerProperty(layoutCyclePosition);
         this.customPreviewLayout = new SimpleObjectProperty<>(customPreviewLayout);
@@ -40,8 +47,17 @@ public class PreviewPreferences {
         this.shouldDownloadCovers.set(shouldDownloadCovers);
     }
 
-    private PreviewPreferences() {
-        this(List.of(), 0, new TextBasedPreviewLayout("", null, null), "", false, false, List.of(), false);
+    private PreviewPreferences(){
+        this(
+                List.of(),
+                0,
+                new TextBasedPreviewLayout("", null, null),
+                "",
+                false,
+                false,
+                List.of(),
+                false
+        );
     }
 
     public static PreviewPreferences getDefault() {
@@ -73,7 +89,9 @@ public class PreviewPreferences {
     }
 
     public PreviewLayout getSelectedPreviewLayout() {
-        if (layoutCycle.isEmpty() || layoutCyclePosition.getValue() < 0 || layoutCyclePosition.getValue() >= layoutCycle.size()) {
+        if (layoutCycle.isEmpty()
+                || layoutCyclePosition.getValue() < 0
+                || layoutCyclePosition.getValue() >= layoutCycle.size()) {
             return getCustomPreviewLayout();
         } else {
             return layoutCycle.get(layoutCyclePosition.getValue());
@@ -128,7 +146,7 @@ public class PreviewPreferences {
         this.bstPreviewLayoutPaths.setAll(bstPreviewLayoutPaths);
     }
 
-    public void setAll(PreviewPreferences preferences) {
+    public void setAll(PreviewPreferences preferences){
         this.layoutCycle.setAll(preferences.getLayoutCycle());
         this.layoutCyclePosition.set(preferences.getLayoutCyclePosition());
         this.customPreviewLayout.set(preferences.getCustomPreviewLayout());